py-server/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-api-python-client==2.107.0
youtube-transcript-api==0.6.1
google-generativeai==0.3.1
markdown==3.5.1
